pyproject.toml
[tool.poetry]
name = "ingestion-api-gcs"
version = "0.1.0"
description = ""
authors = ["Marcus Holanda <mvsh777@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^8.3.3"
requests = "^2.32.3"
google-cloud-storage = "^2.18.2"
google-cloud-secret-manager = "^2.20.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
